// src/sdk/ws-events.ts
// WebSocket Event Type Definitions

import { OperationEntity, PipelineDto } from '../autogenerated/schemas';

export interface BaseWsEvent<T> {
    timestamp: string;
    data: T;
}

export interface OperationEvent extends BaseWsEvent<OperationEntity> {}
export interface PipelineEvent extends BaseWsEvent<PipelineDto> {}

export interface WsEventMap {
    'operation:created': OperationEvent;
    'operation:updated': OperationEvent;
    'pipeline:created': PipelineEvent;
    'pipeline:updated': PipelineEvent;
}